Add files via upload
This dynamic web application is built with a powerful combination of HTML, CSS, and JavaScript, offering a seamless user experience. Leveraging the Ninja API, we fetch up-to-the-minute weather data to provide you with accurate and real-time weather updates for your location or any place you want to explore. With a clean and user-friendly interface, you can effortlessly check current conditions .